pyproject.toml
[project]
name = "agent2agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
]
